fix: updating token icons to production hostname rawcdn.githack.com
As per http://raw.githack.com/:

rawcdn.githack.com is the production hostname providing additional benefits (caching improvements):

No traffic limits or throttling. Files are served via CloudFlare's CDN.
Files can be automatically optimized if you add ?min=1 query parameter.
Use a specific tag or commit hash in the URL (not a branch). Files are cached permanently based on the URL. Query strings are ignored.
The catch: this is a free service, so there are no uptime or support guarantees.